+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * 2.2.0
 
 * Add filesize and mimetype to successful download notification (#48)
+* Add experimental click-to-save feature (#15, #20)
 
 # 2.1.0